Разработать api которое считает количество вопросов в базе #40

+++ src/main/java/com/javamentor/qa/platform/dao/abstracts/model/QuestionDao.java
@@ -6,5 +6,5 @@ import com.javamentor.qa.platform.models.entity.question.Question;
 import java.util.Optional;
 
 public interface QuestionDao extends ReadWriteDao<Question, Long> {
-    Optional<Integer> getCountQuestion();
+    Optional<Long> getCountQuestion();
 }
